Rename C Programming Assignment 1 to C Programming Assignment 1.0
This is a C program that asks the user to create a whole (integer) number, displaying the given input in integer (normal %d), octal (base-8 %o) and hexadecimal (base-16 %x) form. This program also asks the user to enter a real (float) number displaying the given input in float (normal %f) and scientific notion (%e) form and lastly asks the user to enter a character (char) displaying the given input in char (normal %c) and integer (%d) form.